[Fix] #56 - String extension 수정
